command done ai statemachine otw

+++ Assets/Script/Kodingan Elshad/SO/ScriptSO/WeaponStatSO.cs
@@ -14,11 +14,11 @@ public class WeaponStatSO : ScriptableObject
     public float legDamageMultiplier;       //damage yang diterima di kaki
     public float recoil;                    //recoil tembakan nya
     public float reloadTime;                //waktu yang diperlukan untuk reload
-    public int range;                       //jarak yang bisa dipakai oleh tembakan nya
-    public int magSize;                     //jumlah peluru yang bisa di pakai dalam 1 magazine
-    public int magSpare;                    //jumlah magazine yang dibawa
-    public int totalBullet;                    //jumlah magazine yang dibawa
-    public int currBullet;                    //jumlah magazine yang dibawa
+    public float range;                       //jarak yang bisa dipakai oleh tembakan nya
+    public float magSize;                     //jumlah peluru yang bisa di pakai dalam 1 magazine
+    public float magSpare;                    //jumlah magazine yang dibawa
+    public float totalBullet;                    //jumlah magazine yang dibawa
+    public float currBullet;                    //jumlah magazine yang dibawa
     public bool allowHoldDownButton;        //true jika senjatanya automatic
                                             
     public void Initialise()
